requirements.txt
pytest==3.7.4
requests==2.19.1
vcrpy==1.13.0
pytest_rerunfailures==4.1
pytest-xdist==1.23.0
selenium==3.14.1
pytest-selenium==1.13.0
jira
webdriver-manager==1.7
allure_pytest==2.5.2
allure_python_commons==2.5.2